refactor ddl.py and transform.py
